updating prod with master (#808)
* Bump com.microsoft.graph:microsoft-graph in /apps/backend

Bumps [com.microsoft.graph:microsoft-graph](https://github.com/microsoftgraph/msgraph-sdk-java) from 5.77.0 to 5.79.0.
- [Release notes](https://github.com/microsoftgraph/msgraph-sdk-java/releases)
- [Changelog](https://github.com/microsoftgraph/msgraph-sdk-java/blob/dev/CHANGELOG.md)
- [Commits](https://github.com/microsoftgraph/msgraph-sdk-java/compare/v5.77.0...v5.79.0)

---
updated-dependencies:
- dependency-name: com.microsoft.graph:microsoft-graph
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump prettier from 3.2.3 to 3.2.4 in /apps/frontend

Bumps [prettier](https://github.com/prettier/prettier) from 3.2.3 to 3.2.4.
- [Release notes](https://github.com/prettier/prettier/releases)
- [Changelog](https://github.com/prettier/prettier/blob/main/CHANGELOG.md)
- [Commits](https://github.com/prettier/prettier/compare/3.2.3...3.2.4)

---
updated-dependencies:
- dependency-name: prettier
  dependency-type: direct:development
  update-type: version-update:semver-patch
...


* Bump actions/cache from 3 to 4

Bumps [actions/cache](https://github.com/actions/cache) from 3 to 4.
- [Release notes](https://github.com/actions/cache/releases)
- [Changelog](https://github.com/actions/cache/blob/main/RELEASES.md)
- [Commits](https://github.com/actions/cache/compare/v3...v4)

---
updated-dependencies:
- dependency-name: actions/cache
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Oppdatere dokumentere etterlevelse (#801)

* DokumentereEtterlevelse: Bytter ut med aksel komponenter


* DokumentereEtterlevelse: Bytter ut med aksel komponenter del 2


* Fikset tabs til egne komponenter


* Fikset tabs med riktig navn


* Erstatte gamle tabs med nye woot woot


* UPDATE: to use aksel components del 3


* Flyttet EtterlevelseDokumentasjonerPa
nel til egen komponent


* Gjort om EtterlevelseDokumentasjonsPanel til en egen komponent


* Byttet til Aksel komponenter, søk fungerer ikke


* Fikset query feil med riktig query, funger med søk

* Byttet til Aksel komponenter


* Gjorde koden mer lesbar


* UPDATE: component to stop using base ui

* UPDATE: component to stop using base ui

* ADD: text for when there is no dokumentation for team

* UPDATE: to secondary variant

* ListPageHeader til en egen komponent


---------


* Prod (#805)

* Bump com.microsoft.graph:microsoft-graph in /apps/backend

Bumps [com.microsoft.graph:microsoft-graph](https://github.com/microsoftgraph/msgraph-sdk-java) from 5.77.0 to 5.79.0.
- [Release notes](https://github.com/microsoftgraph/msgraph-sdk-java/releases)
- [Changelog](https://github.com/microsoftgraph/msgraph-sdk-java/blob/dev/CHANGELOG.md)
- [Commits](https://github.com/microsoftgraph/msgraph-sdk-java/compare/v5.77.0...v5.79.0)

---
updated-dependencies:
- dependency-name: com.microsoft.graph:microsoft-graph
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump actions/cache from 3 to 4

Bumps [actions/cache](https://github.com/actions/cache) from 3 to 4.
- [Release notes](https://github.com/actions/cache/releases)
- [Changelog](https://github.com/actions/cache/blob/main/RELEASES.md)
- [Commits](https://github.com/actions/cache/compare/v3...v4)

---
updated-dependencies:
- dependency-name: actions/cache
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Bump prettier from 3.2.3 to 3.2.4 in /apps/frontend

Bumps [prettier](https://github.com/prettier/prettier) from 3.2.3 to 3.2.4.
- [Release notes](https://github.com/prettier/prettier/releases)
- [Changelog](https://github.com/prettier/prettier/blob/main/CHANGELOG.md)
- [Commits](https://github.com/prettier/prettier/compare/3.2.3...3.2.4)

---
updated-dependencies:
- dependency-name: prettier
  dependency-type: direct:development
  update-type: version-update:semver-patch
...


* Oppdatere dokumentere etterlevelse (#801)

* DokumentereEtterlevelse: Bytter ut med aksel komponenter


* DokumentereEtterlevelse: Bytter ut med aksel komponenter del 2


* Fikset tabs til egne komponenter


* Fikset tabs med riktig navn


* Erstatte gamle tabs med nye woot woot


* UPDATE: to use aksel components del 3


* Flyttet EtterlevelseDokumentasjonerPa
nel til egen komponent


* Gjort om EtterlevelseDokumentasjonsPanel til en egen komponent


* Byttet til Aksel komponenter, søk fungerer ikke


* Fikset query feil med riktig query, funger med søk

* Byttet til Aksel komponenter


* Gjorde koden mer lesbar


* UPDATE: component to stop using base ui

* UPDATE: component to stop using base ui

* ADD: text for when there is no dokumentation for team

* UPDATE: to secondary variant

* ListPageHeader til en egen komponent


---------


---------


* REFACTOR: method parameters and readability


* REFACTOR: moved component to own file


* UPDATE: tags for new krav/version to show when status is oppfylles senere


* WIP: KravCard refaktorering + ny status (#806)

* KravCard refaktorering + ny status


* MOVED: query to own file

* MOVED: kravfilter interface to constants

* MOVED: queries to own file

* fix: lint warning

* REFACTOR: moved component to own file

* Fikset variabelnavn med query på slutten


* Fikset sortering for både Firefox og Chrome


* Lagt på sjekk for ny versjon


---------


* Fikset lint feil


* Gjort koden mer lesbar


* Laget en egen komponent og gjort koden mer lesbar


* UPDATE: files to be able to run locally and added guide in readme on how to run locally

---------
