+++ apps/frontend/src/api/KravApi.ts
@@ -1,4 +1,3 @@
-import { gql } from '@apollo/client'
 import axios from 'axios'
 import { useEffect, useState } from 'react'
 import { EKravStatus, IKrav, IPageResponse, TKravQL, TOr, emptyPage } from '../constants'
@@ -234,329 +233,3 @@ export const kravMapToFormVal = (krav: Partial<TKravQL>): TKravQL => ({
   aktivertDato: krav.aktivertDato || '',
   prioriteringsId: '',
 })
-
-export const kravFullQuery = gql`
-  query getKrav($id: ID, $kravNummer: Int, $kravVersjon: Int) {
-    kravById(id: $id, nummer: $kravNummer, versjon: $kravVersjon) {
-      id
-      kravNummer
-      kravVersjon
-
-      navn
-      beskrivelse
-      hensikt
-      utdypendeBeskrivelse
-      versjonEndringer
-
-      dokumentasjon
-      implementasjoner
-      kravIdRelasjoner
-      kravRelasjoner {
-        id
-        kravNummer
-        kravVersjon
-        navn
-      }
-      begrepIder
-      begreper {
-        id
-        navn
-        beskrivelse
-      }
-      changeStamp {
-        lastModifiedBy
-        lastModifiedDate
-      }
-      varslingsadresser {
-        adresse
-        type
-        slackChannel {
-          id
-          name
-          numMembers
-        }
-        slackUser {
-          id
-          name
-        }
-      }
-      rettskilder
-      tagger
-      regelverk {
-        lov {
-          code
-          shortName
-        }
-        spesifisering
-      }
-
-      avdeling {
-        code
-        shortName
-      }
-      underavdeling {
-        code
-        shortName
-      }
-      relevansFor {
-        code
-        shortName
-      }
-      suksesskriterier {
-        id
-        navn
-        beskrivelse
-        behovForBegrunnelse
-      }
-      status
-      aktivertDato
-    }
-  }
-`
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-export const etterlevelseDokumentasjonKravQuery = gql`
-  query getKravByFilter(
-    $etterlevelseDokumentasjonId: String
-    $lover: [String!]
-    $gjeldendeKrav: Boolean
-    $etterlevelseDokumentasjonIrrevantKrav: Boolean
-    $status: [String!]
-  ) {
-    krav(
-      filter: {
-        etterlevelseDokumentasjonId: $etterlevelseDokumentasjonId
-        lover: $lover
-        gjeldendeKrav: $gjeldendeKrav
-        etterlevelseDokumentasjonIrrevantKrav: $etterlevelseDokumentasjonIrrevantKrav
-        status: $status
-      }
-    ) {
-      content {
-        id
-        navn
-        kravNummer
-        kravVersjon
-        varselMelding
-        status
-        aktivertDato
-        kravIdRelasjoner
-        prioriteringsId
-        kravRelasjoner {
-          id
-          kravNummer
-          kravVersjon
-          navn
-        }
-        suksesskriterier {
-          id
-          navn
-          beskrivelse
-        }
-        relevansFor {
-          code
-        }
-        regelverk {
-          lov {
-            code
-            shortName
-          }
-        }
-        changeStamp {
-          lastModifiedBy
-          lastModifiedDate
-          createdDate
-        }
-        etterlevelser(onlyForEtterlevelseDokumentasjon: true) {
-          id
-          etterleves
-          fristForFerdigstillelse
-          status
-          changeStamp {
-            lastModifiedBy
-            lastModifiedDate
-          }
-        }
-      }
-    }
-  }
-`
-
-export const KravMedPrioriteringOgEtterlevelseQuery = gql`
-  query getKravByFilter(
-    $etterlevelseDokumentasjonId: String
-    $lover: [String!]
-    $gjeldendeKrav: Boolean
-    $etterlevelseDokumentasjonIrrevantKrav: Boolean
-    $status: [String!]
-  ) {
-    krav(
-      filter: {
-        etterlevelseDokumentasjonId: $etterlevelseDokumentasjonId
-        lover: $lover
-        gjeldendeKrav: $gjeldendeKrav
-        etterlevelseDokumentasjonIrrevantKrav: $etterlevelseDokumentasjonIrrevantKrav
-        status: $status
-      }
-    ) {
-      content {
-        navn
-        kravNummer
-        kravVersjon
-        status
-        prioriteringsId
-        etterlevelser(onlyForEtterlevelseDokumentasjon: true) {
-          id
-          etterleves
-          fristForFerdigstillelse
-          status
-          changeStamp {
-            lastModifiedBy
-            lastModifiedDate
-          }
-        }
-      }
-    }
-  }
-`
-// eslint-enable-next-line @typescript-eslint/ban-types
-
-export const statsQuery = gql`
-  query getBehandlingStats($behandlingId: ID) {
-    behandling(filter: { id: $behandlingId }) {
-      content {
-        stats {
-          fyltKrav {
-            kravNummer
-            kravVersjon
-            status
-            aktivertDato
-            kravIdRelasjoner
-            kravRelasjoner {
-              id
-              kravNummer
-              kravVersjon
-              navn
-            }
-            etterlevelser(onlyForBehandling: true) {
-              behandlingId
-              status
-            }
-            regelverk {
-              lov {
-                code
-                shortName
-              }
-            }
-            changeStamp {
-              lastModifiedBy
-              lastModifiedDate
-              createdDate
-            }
-          }
-          ikkeFyltKrav {
-            kravNummer
-            kravVersjon
-            status
-            aktivertDato
-            kravIdRelasjoner
-            kravRelasjoner {
-              id
-              kravNummer
-              kravVersjon
-              navn
-            }
-            etterlevelser(onlyForBehandling: true) {
-              behandlingId
-              status
-            }
-            regelverk {
-              lov {
-                code
-                shortName
-              }
-            }
-            changeStamp {
-              lastModifiedBy
-              lastModifiedDate
-              createdDate
-            }
-          }
-          irrelevantKrav {
-            kravNummer
-            kravVersjon
-            status
-            aktivertDato
-            kravIdRelasjoner
-            kravRelasjoner {
-              id
-              kravNummer
-              kravVersjon
-              navn
-            }
-            etterlevelser(onlyForBehandling: true) {
-              behandlingId
-              status
-            }
-            regelverk {
-              lov {
-                code
-                shortName
-              }
-            }
-            changeStamp {
-              lastModifiedBy
-              lastModifiedDate
-              createdDate
-            }
-          }
-          lovStats {
-            lovCode {
-              code
-              shortName
-            }
-            fyltKrav {
-              id
-              kravNummer
-              kravVersjon
-              status
-              navn
-              kravIdRelasjoner
-              kravRelasjoner {
-                id
-                kravNummer
-                kravVersjon
-                navn
-              }
-              changeStamp {
-                lastModifiedBy
-                lastModifiedDate
-                createdDate
-              }
-            }
-            ikkeFyltKrav {
-              id
-              kravNummer
-              kravVersjon
-              status
-              aktivertDato
-              navn
-              kravIdRelasjoner
-              kravRelasjoner {
-                id
-                kravNummer
-                kravVersjon
-                navn
-              }
-              changeStamp {
-                lastModifiedBy
-                lastModifiedDate
-                createdDate
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`
