+++ apps/frontend/src/components/Chart.tsx
@@ -1,5 +1,7 @@
 /* eslint-disable jsx-a11y/mouse-events-have-key-events */
+
 /* eslint-disable jsx-a11y/no-static-element-interactions */
+
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import { faChartBar, faChartPie, faCircle } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
