+++ apps/frontend/src/pages/MainPage.tsx
@@ -15,9 +15,10 @@ import {
   IPageResponse,
   TEtterlevelseDokumentasjonQL,
 } from '../constants'
+import { getEtterlevelseDokumentasjonListQuery } from '../query/EtterlevelseDokumentasjonQuery'
 import { ampli, userRoleEventProp } from '../services/Amplitude'
 import { user } from '../services/User'
-import { TVariables, query } from './MyEtterlevelseDokumentasjonerPage'
+import { TVariables } from './MyEtterlevelseDokumentasjonerPage'
 
 export const MainPage = () => {
   const [forsideVarsel, setForsideVarsle] = useState<IMelding>()
@@ -25,7 +26,7 @@ export const MainPage = () => {
   const { data, loading: etterlevelseDokumentasjonLoading } = useQuery<
     { etterlevelseDokumentasjoner: IPageResponse<TEtterlevelseDokumentasjonQL> },
     TVariables
-  >(query, {
+  >(getEtterlevelseDokumentasjonListQuery, {
     variables: { sistRedigert: 4 },
     skip: !user.isLoggedIn(),
   })
