+++ apps/frontend/src/pages/MyEtterlevelseDokumentasjonerPage.tsx
@@ -1,4 +1,3 @@
-import { gql } from '@apollo/client'
 import EditEtterlevelseDokumentasjonModal from '../components/etterlevelseDokumentasjon/edit/EditEtterlevelseDokumentasjonModal'
 import { DokumentasjonTabs } from '../components/etterlevelseDokumentasjon/tabs/DokumentasjonsTabs'
 import { ListPageHeader } from '../components/scaffold/ListPageHeader'
@@ -49,43 +48,3 @@ export type TVariables = {
   teams?: string[]
   behandlingId?: string
 }
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-export const query = gql`
-  query getEtterlevelseDokumentasjoner(
-    $pageNumber: NonNegativeInt
-    $pageSize: NonNegativeInt
-    $mineEtterlevelseDokumentasjoner: Boolean
-    $sistRedigert: NonNegativeInt
-    $sok: String
-    $behandlingId: String
-  ) {
-    etterlevelseDokumentasjoner: etterlevelseDokumentasjon(
-      filter: {
-        mineEtterlevelseDokumentasjoner: $mineEtterlevelseDokumentasjoner
-        sistRedigert: $sistRedigert
-        sok: $sok
-        behandlingId: $behandlingId
-      }
-      pageNumber: $pageNumber
-      pageSize: $pageSize
-    ) {
-      pageNumber
-      pageSize
-      pages
-      numberOfElements
-      totalElements
-      content {
-        id
-        title
-        etterlevelseNummer
-        sistEndretEtterlevelse
-        teamsData {
-          id
-          name
-        }
-      }
-    }
-  }
-`
-// eslint-enable-next-line @typescript-eslint/ban-types
