+++ apps/frontend/src/util/checkAge.tsx
@@ -1,4 +1,7 @@
-export const getNumberOfDaysBetween = (startDate: Date, endDate: Date) => {
-  const diff = Math.abs(startDate.getTime() - endDate.getTime())
+import moment from 'moment'
+
+export const getNumberOfDaysBetween = (startDate: string, endDate: Date) => {
+  const startDateTime = moment(startDate).toDate().getTime()
+  const diff = Math.abs(startDateTime - endDate.getTime())
   return diff / (1000 * 60 * 60 * 24)
 }
